// src/generated/types.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type CartItemInput = {
  /** An array of entered options for the base product, such as personalization text */
  entered_options?: InputMaybe<Array<InputMaybe<EnteredOptionInput>>>;
  /** For child products, the SKU of its parent product */
  parent_sku?: InputMaybe<Scalars['String']>;
  quantity: Scalars['Float'];
  /** The selected options for the base product, such as color or size with  unique ID for a `CustomizableRadioOption`, `CustomizableDropDownOption`, `ConfigurableProductOptionsValues`, etc. objects */
  selected_options?: InputMaybe<Array<InputMaybe<Scalars['ID']>>>;
  sku: Scalars['String'];
};
// ...
export type ComplexTextValue = {
  __typename?: 'ComplexTextValue';
  /** HTML format */
  html: Scalars['String'];
};
// ...
export type Products = {
  __typename?: 'Products';
  /** Layered navigation aggregations. */
  aggregations?: Maybe<Array<Maybe<Aggregation>>>;
  /**
   * Layered navigation filters array.
   * @deprecated Use aggregations instead
   */
  filters?: Maybe<Array<Maybe<LayerFilter>>>;
  /** An array of products that match the specified search criteria. */
  items?: Maybe<Array<Maybe<ProductInterface>>>;
  /** An object that includes the page_info and currentPage values specified in the query. */
  page_info?: Maybe<SearchResultPageInfo>;
  /** An object that includes the default sort field and all available sort fields. */
  sort_fields?: Maybe<SortFields>;
  /** The number of products that are marked as visible. By default, in complex products, parent products are visible, but their child products are not. */
  total_count?: Maybe<Scalars['Int']>;
};
// ...
export type ProductBundleFragment = {
  __typename?: 'BundleProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};

export type ProductConfigurableFragment = {
  __typename?: 'ConfigurableProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
  configurable_options?:
    | Array<
        | {
            __typename?: 'ConfigurableProductOptions';
            uid: string;
            label?: string | null | undefined;
            position?: number | null | undefined;
            use_default?: boolean | null | undefined;
            attribute_code?: string | null | undefined;
            product_id?: number | null | undefined;
            id: string;
            values?:
              | Array<
                  | {
                      __typename?: 'ConfigurableProductOptionsValues';
                      uid?: string | null | undefined;
                      label?: string | null | undefined;
                      id?: string | null | undefined;
                    }
                  | null
                  | undefined
                >
              | null
              | undefined;
          }
        | null
        | undefined
      >
    | null
    | undefined;
};

export type ProductDownloadableFragment = {
  __typename?: 'DownloadableProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};

export type ProductGroupedFragment = {
  __typename?: 'GroupedProduct';
  uid: string;
  sku?: string | null | undefined;
  url_key?: string | null | undefined;
  url_suffix?: string | null | undefined;
  name?: string | null | undefined;
  id: string;
  description?: { __typename?: 'ComplexTextValue'; html: string } | null | undefined;
  thumbnail?:
    | {
        __typename?: 'ProductImage';
        url?: string | null | undefined;
        label?: string | null | undefined;
        id?: string | null | undefined;
      }
    | null
    | undefined;
  media_gallery?:
    | Array<
        | {
            __typename?: 'ProductImage';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductImage';
          }
        | {
            __typename?: 'ProductVideo';
            label?: string | null | undefined;
            url?: string | null | undefined;
            id?: string | null | undefined;
            type: 'ProductVideo';
          }
        | null
        | undefined
      >
    | null
    | undefined;
  price_range: {
    __typename?: 'PriceRange';
    minimum_price: {
      __typename?: 'ProductPrice';
      discount?:
        | { __typename?: 'ProductDiscount'; amount_off?: number | null | undefined }
        | null
        | undefined;
      regular_price: {
        __typename?: 'Money';
        currency?: CurrencyEnum | null | undefined;
        value?: number | null | undefined;
      };
    };
    maximum_price?:
      | {
          __typename?: 'ProductPrice';
          discount?:
            | {
                __typename?: 'ProductDiscount';
                amount_off?: number | null | undefined;
              }
            | null
            | undefined;
          regular_price: {
            __typename?: 'Money';
            currency?: CurrencyEnum | null | undefined;
            value?: number | null | undefined;
          };
        }
      | null
      | undefined;
  };
};
// ...
export type ProductsQueryVariables = Exact<{
  search?: InputMaybe<Scalars['String']>;
  filters?: InputMaybe<ProductAttributeFilterInput>;
  pageSize?: InputMaybe<Scalars['Int']>;
  page?: InputMaybe<Scalars['Int']>;
}>;
// ...
export const AddProductsToCartDocument = gql`
  mutation AddProductsToCart($cartId: String!, $cartItem: CartItemInput!) {
    addProductsToCart(cartId: $cartId, cartItems: [$cartItem]) {
      cart {
        items {
          product {
            name
            sku
          }
          quantity
        }
        total_quantity
      }
    }
  }
`;
// ...
export const CreateEmptyCartDocument = gql`
  mutation CreateEmptyCart {
    createEmptyCart
  }
`;
// ...
export const ProductsDocument = gql`
  query Products(
    $search: String
    $filters: ProductAttributeFilterInput
    $pageSize: Int = 9
    $page: Int = 1
  ) {
    products(
      search: $search
      filter: $filters
      pageSize: $pageSize
      currentPage: $page
    ) {
      ...products
    }
  }
  ${ProductsFragmentDoc}
`;